Changed supersede_lease() to return 1, even if the commit option is not specified.  (Before, it would always return 0 if commit was not given.)  Fixed up the one call to supersede_lease() which expected the old behavior.
Added a check to supersede_lease() to return an error if the pimmediate
flag is given, and commit is not.  (You should never be sending information
on an uncommitted lease to a peer.)

Separated the failover queue update (the propogate flag) test from
the commit test in supersede_lease(), so that you can now enqueue an
update on an uncommitted lease.