[master] Be more liberal when trying to find a group for a host prefix
When trying to find the group to use as a scope boundary for
a prefix from a host declaration relax the requirement that the
prefix be within a subnet.  We now look for the subnet based on
the prefix, if that fails we look for a subnet based on the host
address if it exists and if that fails we use the one from the
shared network we found for the client.