[master] Add code to parse a vendor option as directed by the user.
Add code to parse a vendor option.  It is structured as an action
in the config language.  When the statement is executed it attempts
to find a vendor option in the packet and a vendor option space
specified by the admin for use with that packet.  It then calls
the proper parse routine to do the parsing.