[master] Update DDNS description a bit
There were still some vestiges of the ad-hoc DDNS style
in the documentation.  Mostly places that said only the
interim style allowed a feature but one place mentioned
ad-hoc as well.  As there is no ad-hoc and both interim
and standard are the same for features I've removed such
language.