[master] Correct the dhcp-cache-threshold codes treatment of hostname
If a lease is reused as part of dhcp-cache-threshold processing
the hostname (if it exists) needs to be copied from the temporary
lease back to the permanent lease.

Also add additional checks to see if the lease can be reused.
The host pointer, client ID and hardware address shouldn't have
changed.